#9 txt/csv/json common methods refactor
